+++ src/main/java/youth/week6/member/controller/MemberController.java
@@ -9,24 +9,27 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import youth.week6.argumentResolver.MemberId;
-import youth.week6.member.dto.request.LoginRequestDto;
-import youth.week6.member.dto.request.OrganizerMemberJoinRequestDto;
-import youth.week6.member.dto.request.ParticipantMemberJoinRequestDto;
-import youth.week6.member.dto.response.MemberDetailResponseDto;
-import youth.week6.member.dto.response.MemberJoinResponseDto;
+import youth.week6.member.controller.dto.request.LoginRequestDto;
+import youth.week6.member.controller.dto.request.OrganizerJoinRequestDto;
+import youth.week6.member.controller.dto.request.OrganizerMemberJoinRequestDto;
+import youth.week6.member.controller.dto.request.ParticipantJoinRequestDto;
+import youth.week6.member.controller.dto.request.ParticipantMemberJoinRequestDto;
+import youth.week6.member.controller.dto.response.MemberDetailResponseDto;
+import youth.week6.member.controller.dto.response.MemberJoinResponseDto;
 import youth.week6.member.jwt.dto.TokenInfo;
 import youth.week6.member.controller.mapper.DtosToMemberDetailResponseDtoMapper;
 import youth.week6.member.controller.mapper.MemberJoinRequestDtoToJoinDtoMapper;
 import youth.week6.member.controller.mapper.OrganizerJoinRequestDtoToJoinDtoMapper;
 import youth.week6.member.controller.mapper.ParticipantJoinRequestDtoToJoinDtoMapper;
-import youth.week6.member.member.dto.MemberJoinDto;
-import youth.week6.member.organizer.dto.OrganizerJoinDto;
-import youth.week6.member.participant.dto.ParticipantJoinDto;
+import youth.week6.member.dto.MemberJoinDto;
+import youth.week6.member.dto.OrganizerJoinDto;
+import youth.week6.member.dto.ParticipantJoinDto;
 import youth.week6.member.service.LoginFacadeService;
 import youth.week6.member.service.MemberFacadeService;
 import youth.week6.member.service.dto.MemberDetailDto;
@@ -123,5 +126,49 @@ public class MemberController {
         );
 
         return ResponseEntity.ok(memberDetailResponseDto);
+    }
+
+    @PatchMapping("/participants")
+    public ResponseEntity<?> addRole(
+        @Validated @RequestBody ParticipantJoinRequestDto participantJoinRequestDto,
+        BindingResult bindingResult,
+        @MemberId Long memberId
+    ) {
+
+        // TODO: 2023/06/01 bindingResult Error Handling
+        if (bindingResult.hasErrors()) {
+            Map<String, String> errors = new HashMap<>();
+            bindingResult.getAllErrors()
+                .forEach(c -> errors.put(((FieldError) c).getField(), c.getDefaultMessage()));
+            return ResponseEntity.badRequest().body(errors);
+        }
+
+        ParticipantJoinDto participantJoinDto = participantJoinDtoMapper.to(participantJoinRequestDto);
+
+        memberFacadeService.joinParticipant(memberId, participantJoinDto);
+
+        return ResponseEntity.ok().build();
+    }
+
+    @PatchMapping("/organizers")
+    public ResponseEntity<?> addRole(
+        @Validated @RequestBody OrganizerJoinRequestDto organizerJoinRequestDto,
+        BindingResult bindingResult,
+        @MemberId Long memberId
+    ) {
+
+        // TODO: 2023/06/01 bindingResult Error Handling
+        if (bindingResult.hasErrors()) {
+            Map<String, String> errors = new HashMap<>();
+            bindingResult.getAllErrors()
+                .forEach(c -> errors.put(((FieldError) c).getField(), c.getDefaultMessage()));
+            return ResponseEntity.badRequest().body(errors);
+        }
+
+        OrganizerJoinDto organizerJoinDto = organizerJoinDtoMapper.to(organizerJoinRequestDto);
+
+        memberFacadeService.joinOrganizer(memberId, organizerJoinDto);
+
+        return ResponseEntity.ok().build();
     }
 }