+++ src/test/java/youth/week6/member/controller/MemberControllerTest.java
@@ -6,13 +6,19 @@ import static youth.week6.testFixture.AllergensFixture.MILK;
 import static youth.week6.testFixture.MemberFixture.사용자정보_정상입력;
 import static youth.week6.testFixture.OrganizerFixture.주최자정보_정상입력;
 import static youth.week6.testFixture.ParticipantFixture.참여자정보_정상입력;
+import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.주최자권한_추가됨;
+import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.주최자권한_추가불가;
 import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.주최자정보_확인됨;
 import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.참여자_회원가입_확인됨;
+import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.참여자권한_추가됨;
+import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.참여자권한_추가불가;
 import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.참여자정보_확인됨;
 import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.회원가입_중복지원_에러발생;
 import static youth.week6.testUtils.acceptiontesetUtils.assertionUtils.MemberAcceptanceTestAssertionUtils.회원정보_확인됨;
 import static youth.week6.testUtils.acceptiontesetUtils.sendUtils.MemberAcceptanceTestSendUtils.로그인을통한_JWT토큰획득;
+import static youth.week6.testUtils.acceptiontesetUtils.sendUtils.MemberAcceptanceTestSendUtils.주최자_권한_요청;
 import static youth.week6.testUtils.acceptiontesetUtils.sendUtils.MemberAcceptanceTestSendUtils.주최자_회원가입_요청;
+import static youth.week6.testUtils.acceptiontesetUtils.sendUtils.MemberAcceptanceTestSendUtils.참여자_권한_요청;
 import static youth.week6.testUtils.acceptiontesetUtils.sendUtils.MemberAcceptanceTestSendUtils.참여자_회원가입_요청;
 import static youth.week6.testUtils.acceptiontesetUtils.sendUtils.MemberAcceptanceTestSendUtils.회원정보_조회요청;
 
@@ -21,12 +27,12 @@ import io.restassured.response.Response;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import youth.week6.member.dto.request.LoginRequestDto;
-import youth.week6.member.dto.request.MemberJoinRequestDto;
-import youth.week6.member.dto.request.OrganizerJoinRequestDto;
-import youth.week6.member.dto.request.OrganizerMemberJoinRequestDto;
-import youth.week6.member.dto.request.ParticipantJoinRequestDto;
-import youth.week6.member.dto.request.ParticipantMemberJoinRequestDto;
+import youth.week6.member.controller.dto.request.LoginRequestDto;
+import youth.week6.member.controller.dto.request.MemberJoinRequestDto;
+import youth.week6.member.controller.dto.request.OrganizerJoinRequestDto;
+import youth.week6.member.controller.dto.request.OrganizerMemberJoinRequestDto;
+import youth.week6.member.controller.dto.request.ParticipantJoinRequestDto;
+import youth.week6.member.controller.dto.request.ParticipantMemberJoinRequestDto;
 import youth.week6.member.participant.entity.Allergens;
 import youth.week6.testFixture.AllergensFixture;
 import youth.week6.testUtils.SpringBootTestHelper;
@@ -86,11 +92,7 @@ class MemberControllerTest extends SpringBootTestHelper {
     @Test
     public void 참여자_정보조회 (){
         //given
-        ParticipantMemberJoinRequestDto 요청전문 = new ParticipantMemberJoinRequestDto(
-            사용자_정보, 참여자_정보);
-        참여자_회원가입_요청(요청전문);
-        String JWT_토큰 = 로그인을통한_JWT토큰획득(
-            new LoginRequestDto(사용자_정보.getIdentification(), 사용자_정보.getPassword()));
+        String JWT_토큰 = 참여자_회원가입_JWT_토큰받기();
 
         //when
         ExtractableResponse<Response> 회원정보_조회요청_response = 회원정보_조회요청(JWT_토큰);
@@ -102,14 +104,12 @@ class MemberControllerTest extends SpringBootTestHelper {
         );
     }
 
+
+
     @Test
     public void 주최자_정보조회 (){
         //given
-        OrganizerMemberJoinRequestDto 요청전문 = new OrganizerMemberJoinRequestDto(
-            사용자_정보, 주최자_정보);
-        주최자_회원가입_요청(요청전문);
-        String JWT_토큰 = 로그인을통한_JWT토큰획득(
-            new LoginRequestDto(사용자_정보.getIdentification(), 사용자_정보.getPassword()));
+        String JWT_토큰 = 주최자_회원가입_JWT_토큰받기();
 
         //when
         ExtractableResponse<Response> 회원정보_조회요청_response = 회원정보_조회요청(JWT_토큰);
@@ -136,4 +136,68 @@ class MemberControllerTest extends SpringBootTestHelper {
         회원가입_중복지원_에러발생(주최자_회원가입_요청_response);
     }
 
+    @Test
+    public void 참여자권한_신청 (){
+        //given
+        String JWT토큰 = 주최자_회원가입_JWT_토큰받기();
+
+        //when
+        ExtractableResponse<Response> 참여자_권한_요청_response = 참여자_권한_요청(JWT토큰, 참여자_정보);
+
+        //then
+        참여자권한_추가됨(참여자_권한_요청_response);
+    }
+
+    @Test
+    public void 참여자권한_중복신청불가 (){
+        //given
+        String JWT토큰 = 참여자_회원가입_JWT_토큰받기();
+
+        //when
+        ExtractableResponse<Response> 참여자_권한_요청_response = 참여자_권한_요청(JWT토큰, 참여자_정보);
+
+        //then
+        참여자권한_추가불가(참여자_권한_요청_response);
+    }
+
+    @Test
+    public void 주최자권한_신청 (){
+        //given
+        String JWT토큰 = 참여자_회원가입_JWT_토큰받기();
+
+        //when
+        ExtractableResponse<Response> 주최자_권한_요청_response = 주최자_권한_요청(JWT토큰, 주최자_정보);
+
+        //then
+        주최자권한_추가됨(주최자_권한_요청_response);
+    }
+
+    @Test
+    public void 주최자권한_중복신청불가 (){
+        //given
+        String JWT토큰 = 주최자_회원가입_JWT_토큰받기();
+
+        //when
+        ExtractableResponse<Response> 주최자_권한_요청_response = 주최자_권한_요청(JWT토큰, 주최자_정보);
+
+        //then
+        주최자권한_추가불가(주최자_권한_요청_response);
+    }
+
+    private String 주최자_회원가입_JWT_토큰받기() {
+        OrganizerMemberJoinRequestDto 요청전문 = new OrganizerMemberJoinRequestDto(
+            사용자_정보, 주최자_정보);
+        주최자_회원가입_요청(요청전문);
+        return 로그인을통한_JWT토큰획득(
+            new LoginRequestDto(사용자_정보.getIdentification(), 사용자_정보.getPassword()));
+    }
+
+    private String 참여자_회원가입_JWT_토큰받기() {
+        ParticipantMemberJoinRequestDto 요청전문 = new ParticipantMemberJoinRequestDto(
+            사용자_정보, 참여자_정보);
+        참여자_회원가입_요청(요청전문);
+        String JWT_토큰 = 로그인을통한_JWT토큰획득(
+            new LoginRequestDto(사용자_정보.getIdentification(), 사용자_정보.getPassword()));
+        return JWT_토큰;
+    }
 }