+++ readmeSource/HISTORY.md
@@ -34,3 +34,15 @@
      - Argument Resolver를 통해 획득한다
   2. 사용자 ID 기반 사용자 정보, 참여자 정보, 주최자 정보를 가져온다
      - MemberFacadeService를 통해 접근하며, 순서대로 사용자 정보, 참여자 정보, 주최자 정보를 가져온다. 
+
+### add_role
+- 주최자 권한 추가 기능을 구현한다. 
+  1. 이미 주최자 권한이 있다면 추가 로직을 중단한다. 
+  2. Authorization Header 내 Jwt Token을 통해 사용자 ID를 조회한다
+    - Argument Resolver를 통해 획득한다
+  3. 주최자 권한에 필요한 추가정보를 받아 주최자 정보와 Role을 추가한다. 
+- 참여자 권한 추가 기능을 구현한다. 
+  1. 이미 참여자 권한이 있다면 추가 로직을 중단한다. 
+  2. Authorization Header 내 Jwt Token을 통해 사용자 ID를 조회한다
+     - Argument Resolver를 통해 획득한다
+  3. 참여자 권한에 필요한 추가정보를 받아 참여자 정보와 Role을 추가한다.